+++ src/commands/__init__.py
@@ -1,6 +1,7 @@
 from commands import duality_commands as duality_commands
 from commands import error_commands as error_commands
 from commands import timer_commands as timer_commands
+from commands.campaign_commands import master, player, session_start, session_stop
 from commands.duality_commands import duality
 from commands.error_commands import handle_error
 from commands.game_commands import game, game_url
@@ -17,6 +18,8 @@ __all__ = (
     'handle_error',
     'game',
     'game_url',
+    'master',
+    'player',
     'HELP_MESSAGE',
     'COMMANDS',
     'observed_callback',
@@ -24,6 +27,8 @@ __all__ = (
     'roll',
     'roll20',
     'rolld20',
+    'session_start',
+    'session_stop',
     'timer',
     'version_command',
 )